% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{draw.survival}
\alias{draw.survival}
\title{Title draw.survival}
\usage{
draw.survival(
  survfit = fit_PFI,
  ylab = "BCR",
  legend.position = "right",
  plot.title = NULL,
  break.time = NA,
  legend.labs = c("NE.sig High", "NE.sig Low"),
  xlimt = NA,
  risk.table = F,
  front.size = 10,
  palette = c("#1F78B4", "#E31A1C"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title draw.survival
}
