% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_strip_map.R
\name{make_strip_map}
\alias{make_strip_map}
\title{Make strip map plot}
\usage{
make_strip_map(df)
}
\arguments{
\item{df}{a dataframe generated from make_flight_path}
}
\value{
plot
}
\description{
Create a strip map plot
}
