% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{NEPal_bulk}
\alias{NEPal_bulk}
\title{Title NEPal_bulk}
\usage{
NEPal_bulk(
  bulk.data,
  method = c("ssGSEA", "Enet", "Ridge", "SVM", "all"),
  species = c("human", "mouse"),
  gene.sets = c("all", "EurUrol.2005", "CancerDiscov.2011", "CancerRes.2014",
    "NatMed.2016", "CellRep.2018.scRNA", "ClinCancerRes.2015", "JClinInvest.2019",
    "IntJCancer.2019", "HP_NE_neoplasm", "JClinOncol.2018", "BMC.Cancer.2017",
    "NEPal")
)
}
\arguments{
\item{bulk.data}{input bulk transcriptome data.}

\item{gene.sets}{}
}
\description{
Title NEPal_bulk
}
