% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{NEPAL_scRNA}
\alias{NEPAL_scRNA}
\title{Title NEPAL_scRNA}
\usage{
NEPAL_scRNA(
  seurat.data,
  method = "AUCell",
  species = c("human", "mouse"),
  ncores = 2,
  assay.names = "NE",
  DefaultAssay = F,
  ...
)
}
\arguments{
\item{seurat.data}{input scRNA-seq seurat data.}

\item{...}{}
}
\description{
Title NEPAL_scRNA
}
